base_requirements.txt
django-grappelli==3.0.3
django-querysetsequence==0.15
gunicorn==20.1.0
markdown==3.3.6
pip-chill==1.0.1
pre-commit==2.17.0
psycopg2==2.9.3
